requirements.txt
pandas==1.0.4
tqdm==4.49.0
selenium==3.141.0
numpy==1.19.5
webdriver_manager==3.4.2
path.py==12.5.0
progressbar33==2.4
resources==0.0.1
